+++ src/main/java/com/kushagramathur/airbnb_clone/service/impl/RoomServiceImpl.java
@@ -3,7 +3,9 @@ package com.kushagramathur.airbnb_clone.service.impl;
 import com.kushagramathur.airbnb_clone.dto.RoomDto;
 import com.kushagramathur.airbnb_clone.entity.Hotel;
 import com.kushagramathur.airbnb_clone.entity.Room;
+import com.kushagramathur.airbnb_clone.entity.User;
 import com.kushagramathur.airbnb_clone.exception.ResourceNotFoundException;
+import com.kushagramathur.airbnb_clone.exception.UnAuthorisedException;
 import com.kushagramathur.airbnb_clone.repository.HotelRepository;
 import com.kushagramathur.airbnb_clone.repository.InventoryRepository;
 import com.kushagramathur.airbnb_clone.repository.RoomRepository;
@@ -13,6 +15,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -34,6 +37,11 @@ public class RoomServiceImpl implements RoomService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel with id " + hotelId + " not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException(("This user is not the owner of this hotel"));
+        }
 
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
@@ -76,6 +84,11 @@ public class RoomServiceImpl implements RoomService {
         Room room = roomRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Room with id " + id + " not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorisedException(("This user is not the owner of this hotel"));
+        }
 
         inventoryService.deleteFutureInventory(room);
         roomRepository.deleteById(id);
